Brain-tumor-detection/requirement.txt
datasets
ultralytics
matplotlib
Pillow